+++ api/src/main/java/org/allaymc/api/world/WorldPool.java
@@ -1,5 +1,6 @@
 package org.allaymc.api.world;
 
+import lombok.Builder;
 import org.allaymc.api.math.location.Location3i;
 import org.allaymc.api.math.location.Location3ic;
 import org.allaymc.api.world.generator.WorldGenerator;
@@ -63,32 +64,28 @@ public interface WorldPool {
             WorldGenerator netherGenerator,
             WorldGenerator theEndGenerator
     ) {
-        loadWorld(name, storage, overworldGenerator, netherGenerator, theEndGenerator, true, true, true);
+        loadWorld(
+                new WorldSetting(name, storage, false),
+                new DimensionSetting(overworldGenerator, true),
+                new DimensionSetting(netherGenerator, true),
+                new DimensionSetting(theEndGenerator, true)
+        );
     }
 
     /**
-     * Create a world with the provided name, storage, generators and light calculation settings for different dimensions.
+     * Create a world with the provided name, storage, and dimension settings for different dimensions.
      *
-     * @param name                    the name of the world to be loaded
-     * @param storage                 the storage used for reading and writing chunks and world data
-     * @param overworldGenerator      the generator used to initialize or generate the overworld dimension
-     * @param netherGenerator         the generator used to initialize or generate the nether dimension, or
-     *                                {@code null} to disable nether dimension
-     * @param theEndGenerator         the generator used to initialize or generate the end dimension, or
-     *                                {@code null} to disable the end dimension
-     * @param overworldEnableLightCalc whether to enable light calculation for the overworld dimension
-     * @param netherEnableLightCalc    whether to enable light calculation for the nether dimension
-     * @param theEndEnableLightCalc    whether to enable light calculation for the end dimension
+     * @param worldSetting     the setting of the world
+     * @param overworldSetting the setting of the overworld dimension
+     * @param netherSetting    the setting of the nether dimension, or {@code null} if the nether dimension does not exist
+     * @param theEndSetting    the setting of the end, or {@code null} if the end dimension does not exist
      * @throws IllegalArgumentException if the world with the specific name already exists
      */
     void loadWorld(
-            String name, WorldStorage storage,
-            WorldGenerator overworldGenerator,
-            WorldGenerator netherGenerator,
-            WorldGenerator theEndGenerator,
-            boolean overworldEnableLightCalc,
-            boolean netherEnableLightCalc,
-            boolean theEndEnableLightCalc
+            WorldSetting worldSetting,
+            DimensionSetting overworldSetting,
+            DimensionSetting netherSetting,
+            DimensionSetting theEndSetting
     );
 
     /**
@@ -109,5 +106,26 @@ public interface WorldPool {
     default Location3ic getGlobalSpawnPoint() {
         var vec = getDefaultWorld().getWorldData().getSpawnPoint();
         return new Location3i(vec.x(), vec.y(), vec.z(), getDefaultWorld().getOverWorld());
+    }
+
+    /**
+     * WorldSetting contains the setting of a world when creating a new world.
+     *
+     * @param name             the name of the world
+     * @param storage          the storage used for this world
+     * @param useVirtualThread whether to use the virtual thread as the ticking thread of the world and the dimensions in the world
+     */
+    @Builder
+    record WorldSetting(String name, WorldStorage storage, boolean useVirtualThread) {
+    }
+
+    /**
+     * DimensionSetting contains the setting of a dimension when creating a new world.
+     *
+     * @param worldGenerator         the world generator used for this dimension
+     * @param enableLightCalculation whether light calculation is enabled in this dimension
+     */
+    @Builder
+    record DimensionSetting(WorldGenerator worldGenerator, boolean enableLightCalculation) {
     }
 }