+++ server/src/main/java/org/allaymc/server/world/AllayDimension.java
@@ -70,7 +70,7 @@ public class AllayDimension implements Dimension {
     protected final Set<Player> players;
     protected final Set<DebugShape> debugShapes;
 
-    public AllayDimension(AllayWorld world, WorldGenerator worldGenerator, DimensionInfo dimensionInfo, boolean enableLightCalculation) {
+    public AllayDimension(AllayWorld world, WorldGenerator worldGenerator, DimensionInfo dimensionInfo, boolean enableLightCalculation, boolean useVirtualThread) {
         this.world = world;
         this.dimensionInfo = dimensionInfo;
         this.worldGenerator = worldGenerator;
@@ -78,7 +78,7 @@ public class AllayDimension implements Dimension {
         this.entityManager = new AllayEntityManager(this, world.getWorldStorage());
         this.blockUpdateManager = new AllayBlockUpdateManager(this);
         this.scheduler = new AllayScheduler(Server.getInstance().getVirtualThreadPool());
-        this.lightEngine = enableLightCalculation ? new AllayLightEngine(this) : new NoLightEngine(this);
+        this.lightEngine = enableLightCalculation ? new AllayLightEngine(this, useVirtualThread) : new NoLightEngine(this);
         this.players = new NonBlockingHashSet<>();
         this.debugShapes = new NonBlockingHashSet<>();
         worldGenerator.setDimension(this);