+++ server/src/main/java/org/allaymc/server/world/AllayWorldPool.java
@@ -20,7 +20,6 @@ import org.allaymc.api.world.WorldPool;
 import org.allaymc.api.world.WorldState;
 import org.allaymc.api.world.data.DimensionInfo;
 import org.allaymc.api.world.generator.WorldGenerator;
-import org.allaymc.api.world.storage.WorldStorage;
 import org.allaymc.server.utils.Utils;
 import org.allaymc.server.world.light.NoLightEngine;
 
@@ -79,37 +78,31 @@ public final class AllayWorldPool implements WorldPool {
     }
 
     @Override
-    public void loadWorld(
-            String name, WorldStorage storage,
-            WorldGenerator overworldGenerator,
-            WorldGenerator netherGenerator,
-            WorldGenerator theEndGenerator,
-            boolean overworldEnableLightCalc,
-            boolean netherEnableLightCalc,
-            boolean theEndEnableLightCalc
-    ) {
+    public void loadWorld(WorldSetting worldSetting, DimensionSetting overworldSetting, DimensionSetting netherSetting, DimensionSetting theEndSetting) {
+        var name = worldSetting.name();
         log.info(I18n.get().tr(TrKeys.ALLAY_WORLD_LOADING, name));
         if (worlds.containsKey(name)) {
             throw new IllegalArgumentException("World " + name + " is already loaded");
         }
 
+        var useVirtualThread = worldSetting.useVirtualThread();
         AllayWorld world;
         try {
-            world = new AllayWorld(name, storage);
+            world = new AllayWorld(name, worldSetting.storage(), useVirtualThread);
         } catch (Throwable t) {
             log.error("Error while initializing world {}", name, t);
             return;
         }
 
         // Load overworld dimension
-        world.addDimension(new AllayDimension(world, overworldGenerator, DimensionInfo.OVERWORLD, overworldEnableLightCalc));
+        world.addDimension(new AllayDimension(world, overworldSetting.worldGenerator(), DimensionInfo.OVERWORLD, overworldSetting.enableLightCalculation(), useVirtualThread));
 
         // Load nether and the end dimension if they are not null
-        if (netherGenerator != null) {
-            world.addDimension(new AllayDimension(world, netherGenerator, DimensionInfo.NETHER, netherEnableLightCalc));
+        if (netherSetting != null) {
+            world.addDimension(new AllayDimension(world, netherSetting.worldGenerator(), DimensionInfo.NETHER, netherSetting.enableLightCalculation(), useVirtualThread));
         }
-        if (theEndGenerator != null) {
-            world.addDimension(new AllayDimension(world, theEndGenerator, DimensionInfo.THE_END, theEndEnableLightCalc));
+        if (theEndSetting != null) {
+            world.addDimension(new AllayDimension(world, theEndSetting.worldGenerator(), DimensionInfo.THE_END, theEndSetting.enableLightCalculation(), useVirtualThread));
         }
 
         var event = new WorldLoadEvent(world);
@@ -168,11 +161,10 @@ public final class AllayWorldPool implements WorldPool {
         var theEndGenerator = setting.theEnd() == null ? null : tryCreateWorldGenerator(setting.theEnd());
 
         loadWorld(
-                name, storage,
-                overworldGenerator, netherGenerator, theEndGenerator,
-                setting.overworld().enableLightCalculation(),
-                setting.nether() != null ? setting.nether().enableLightCalculation() : true,
-                setting.theEnd() != null ? setting.theEnd().enableLightCalculation() : true
+                new WorldSetting(name, storage, false),
+                new DimensionSetting(overworldGenerator, setting.overworld().enableLightCalculation()),
+                netherGenerator != null ? new DimensionSetting(netherGenerator, setting.nether().enableLightCalculation()) : null,
+                theEndGenerator != null ? new DimensionSetting(theEndGenerator, setting.theEnd().enableLightCalculation()) : null
         );
     }
 
@@ -210,6 +202,7 @@ public final class AllayWorldPool implements WorldPool {
 
         return new WorldSettings.WorldSetting(
                 world.getWorldStorage().getName(),
+                world.isVirtualTickingThread(),
                 new WorldSettings.WorldSetting.DimensionSetting(owGenerator.getName(), owGenerator.getPreset(), overworldLightEnabled),
                 netherGenerator != null ? new WorldSettings.WorldSetting.DimensionSetting(netherGenerator.getName(), netherGenerator.getPreset(), netherLightEnabled) : null,
                 theEndGenerator != null ? new WorldSettings.WorldSetting.DimensionSetting(theEndGenerator.getName(), theEndGenerator.getPreset(), theEndLightEnabled) : null
@@ -223,7 +216,7 @@ public final class AllayWorldPool implements WorldPool {
         @Setter
         @CustomKey("worlds")
         private Map<String, WorldSetting> worlds = Map.of("world", new WorldSetting(
-                "LEVELDB",
+                "LEVELDB", false,
                 new WorldSetting.DimensionSetting(
                         "FLAT",
                         ""
@@ -240,6 +233,10 @@ public final class AllayWorldPool implements WorldPool {
         public static class WorldSetting extends OkaeriConfig {
             @CustomKey("storage-type")
             private String storageType;
+
+            @Comment("If set to true, virtual thread will be used as the ticking thread of the world and the dimensions in the world")
+            @CustomKey("virtual-ticking-thread")
+            private boolean virtualTickingThread;
 
             private DimensionSetting overworld;